Fix IE directive not compiled issue
In the TodoMVC example, when `v-show` is compiled first, it adds an
inline style attribute to the node. Since IE seems to handle the order
of attributes in `node.attributes` differently from other browsers,
this causes some directives to be skipped and never compiled. Simply
copy the attribtues into an Array before compiling solves the issue.